break up selfdrive.config (#23959)
* break up selfdrive.config

* add to release files

* fix import order
