+++ selfdrive/car/ford/radar_interface.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 from cereal import car
+from common.conversions import Conversions as CV
 from opendbc.can.parser import CANParser
 from selfdrive.car.ford.values import DBC
-from selfdrive.config import Conversions as CV
 from selfdrive.car.interfaces import RadarInterfaceBase
 
 RADAR_MSGS = list(range(0x500, 0x540))
