+++ selfdrive/car/gm/radar_interface.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 import math
 from cereal import car
+from common.conversions import Conversions as CV
 from opendbc.can.parser import CANParser
 from selfdrive.car.gm.values import DBC, CAR, CanBus
-from selfdrive.config import Conversions as CV
 from selfdrive.car.interfaces import RadarInterfaceBase
 
 RADAR_HEADER_MSG = 1120
