+++ selfdrive/car/honda/hondacan.py
@@ -1,5 +1,5 @@
+from common.conversions import Conversions as CV
 from selfdrive.car.honda.values import HondaFlags, HONDA_BOSCH, CAR, CarControllerParams
-from selfdrive.config import Conversions as CV
 
 # CAN bus layout with relay
 # 0 = ACC-CAN - radar side
