+++ selfdrive/car/mock/interface.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 import math
 from cereal import car
-from selfdrive.config import Conversions as CV
+from common.conversions import Conversions as CV
 from selfdrive.swaglog import cloudlog
 import cereal.messaging as messaging
 from selfdrive.car import gen_empty_fingerprint, get_safety_config
