+++ tools/replay/lib/ui_helpers.py
@@ -10,8 +10,7 @@ from matplotlib.backends.backend_agg import FigureCanvasAgg
 from common.transformations.camera import (eon_f_frame_size, eon_f_focal_length,
                                            tici_f_frame_size, tici_f_focal_length,
                                            get_view_frame_from_calib_frame)
-from selfdrive.config import UIParams as UP
-from selfdrive.config import RADAR_TO_CAMERA
+from selfdrive.controls.lib.radar_helpers import RADAR_TO_CAMERA
 
 
 RED = (255, 0, 0)
@@ -23,6 +22,15 @@ WHITE = (255, 255, 255)
 
 _FULL_FRAME_SIZE = {
 }
+
+class UIParams:
+  lidar_x, lidar_y, lidar_zoom = 384, 960, 6
+  lidar_car_x, lidar_car_y = lidar_x / 2., lidar_y / 1.1
+  car_hwidth = 1.7272 / 2 * lidar_zoom
+  car_front = 2.6924 * lidar_zoom
+  car_back = 1.8796 * lidar_zoom
+  car_color = 110
+UP = UIParams
 
 _BB_TO_FULL_FRAME = {}
 _CALIB_BB_TO_FULL = {}
