+++ tools/replay/ui.py
@@ -10,8 +10,7 @@ import pygame  # pylint: disable=import-error
 import cereal.messaging as messaging
 from common.numpy_fast import clip
 from common.basedir import BASEDIR
-from selfdrive.config import UIParams as UP
-from tools.replay.lib.ui_helpers import (_BB_TO_FULL_FRAME,
+from tools.replay.lib.ui_helpers import (_BB_TO_FULL_FRAME, UP,
                                          _INTRINSICS, BLACK, GREEN,
                                          YELLOW, Calibration,
                                          get_blank_lid_overlay, init_plots,
